# docs/API-spec.yaml
openapi: 3.0.3
info:
  title: Shopping Food Store API
  version: 0.0.1
paths:
  /api/auth/register:
    post:
      summary: Regista um cliente
      responses:
        "200":
          description: Cliente registado com token JWT
        "400":
          description: Pedido invalido
        "409":
          description: Email duplicado
  /api/auth/login:
    post:
      summary: Autentica um utilizador
      responses:
        "200":
          description: Login com token JWT
        "401":
          description: Credenciais invalidas
  /api/auth/me:
    get:
      summary: Consulta o utilizador autenticado
      responses:
        "200":
          description: Utilizador autenticado
        "401":
          description: Token ausente ou invalido
  /api/categories:
    get:
      summary: Lista categorias
      responses:
        "200":
          description: Lista de categorias
    post:
      summary: Cria uma categoria
      responses:
        "201":
          description: Categoria criada
        "400":
          description: Pedido invalido
        "409":
          description: Categoria duplicada
  /api/categories/{id}:
    get:
      summary: Consulta uma categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Categoria encontrada
        "404":
          description: Categoria nao encontrada
    put:
      summary: Atualiza uma categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Categoria atualizada
        "400":
          description: Pedido invalido
        "404":
          description: Categoria nao encontrada
        "409":
          description: Categoria duplicada
    delete:
      summary: Remove uma categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Categoria removida
        "404":
          description: Categoria nao encontrada
        "409":
          description: Categoria com produtos associados
  /api/products:
    get:
      summary: Lista produtos
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de produtos
    post:
      summary: Cria um produto
      responses:
        "201":
          description: Produto criado
        "400":
          description: Pedido invalido
        "404":
          description: Categoria nao encontrada
  /api/products/{id}:
    get:
      summary: Consulta um produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Produto encontrado
        "404":
          description: Produto nao encontrado
  /api/sales/checkout:
    post:
      summary: Finaliza uma compra
      responses:
        "200":
          description: Venda criada
        "400":
          description: Carrinho invalido ou stock insuficiente
        "401":
          description: Cliente nao autenticado
  /api/sales:
    get:
      summary: Lista vendas do cliente autenticado
      responses:
        "200":
          description: Historico de vendas
        "401":
          description: Cliente nao autenticado
  /api/sales/{id}:
    get:
      summary: Consulta uma venda do cliente autenticado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venda encontrada
        "401":
          description: Cliente nao autenticado
        "404":
          description: Venda nao encontrada
  /api/sales/{id}/invoice:
    get:
      summary: Consulta a fatura de uma venda
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fatura encontrada
        "401":
          description: Cliente nao autenticado
        "404":
          description: Fatura nao encontrada
  /api/admin/sales:
    get:
      summary: Lista todas as vendas
      responses:
        "200":
          description: Listagem global de vendas
        "403":
          description: Acesso reservado a administradores
  /api/invoices/{id}:
    get:
      summary: Consulta uma fatura
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fatura encontrada
        "401":
          description: Cliente nao autenticado
        "404":
          description: Fatura nao encontrada
  /api/stats/products/top-selling:
    get:
      summary: Lista produtos mais vendidos
      responses:
        "200":
          description: Estatisticas de produtos
        "403":
          description: Acesso reservado a administradores
  /api/stats/products/least-selling:
    get:
      summary: Lista produtos menos vendidos
      responses:
        "200":
          description: Estatisticas de produtos
        "403":
          description: Acesso reservado a administradores
  /api/stats/customers/best:
    get:
      summary: Lista melhores clientes
      responses:
        "200":
          description: Estatisticas de clientes
        "403":
          description: Acesso reservado a administradores
  /api/stats/revenue:
    get:
      summary: Consulta faturacao por periodo
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Faturacao do periodo
        "400":
          description: Periodo invalido
        "403":
          description: Acesso reservado a administradores
    put:
      summary: Atualiza um produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Produto atualizado
        "400":
          description: Pedido invalido
        "404":
          description: Produto ou categoria nao encontrado
    delete:
      summary: Remove um produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Produto removido
        "404":
          description: Produto nao encontrado